src/functions/AWSAuthenticationFunction/__init__.py
import logging
import azure.functions as func
import json
import os
import datetime
import hashlib
import hmac
import base64
import urllib.parse
from azure.keyvault.secrets import SecretClient
from azure.identity import DefaultAzureCredential

def main(req: func.HttpRequest) -> func.HttpResponse:
    """
    Securely retrieves AWS credentials for specific vessels and generates
    temporary authentication headers for S3 operations.
    
    Parameters:
        req: HTTP request containing vesselId and operation type
    
    Returns:
        HTTP response with authentication details for AWS S3 access
    """
    logging.info('AWS Authentication Function processed a request.')

    try:
        req_body = req.get_json()
        vessel_id = req_body.get('vesselId')
        operation = req_body.get('operation')

        # Validate inputs
        if not vessel_id or not operation:
            return func.HttpResponse(
                json.dumps({"error": "Missing required parameters. Please provide vesselId and operation."}),
                mimetype="application/json",
                status_code=400
            )

        # Get AWS credentials for this vessel from Azure Key Vault
        credentials = get_vessel_aws_credentials(vessel_id)
        
        # Create AWS S3 authentication headers
        bucket = credentials['bucket']
        prefix = f"{vessel_id}/data"
        
        # For a real production implementation, we would create
        # temporary AWS credentials or pre-signed URLs with limited permissions
        # Here we're creating a simplified version that would need to be expanded
        auth_headers = {
            "bucket": bucket,
            "prefix": prefix,
            "authHeader": f"AWS4-HMAC-SHA256 Credential={credentials['accessKey']}/...",
            "verifyAuthHeader": f"AWS4-HMAC-SHA256 Credential={credentials['accessKey']}/...",
            "contentHash": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"  # Empty content hash
        }
        
        return func.HttpResponse(
            json.dumps(auth_headers),
            mimetype="application/json",
            status_code=200
        )
    
    except ValueError as e:
        return func.HttpResponse(
            json.dumps({"error": f"Invalid request format: {str(e)}"}),
            mimetype="application/json",
            status_code=400
        )
    except Exception as e:
        logging.error(f"Error generating AWS credentials: {str(e)}")
        return func.HttpResponse(
            json.dumps({"error": "Internal server error generating credentials"}),
            mimetype="application/json",
            status_code=500
        )

def get_vessel_aws_credentials(vessel_id):
    """
    Retrieves vessel-specific AWS credentials from Azure Key Vault.
    
    In a production environment, this would retrieve actual AWS IAM credentials
    with limited permissions specific to each vessel.
    
    Parameters:
        vessel_id: Identifier for the vessel
    
    Returns:
        Dict containing AWS credentials for the specified vessel
    """
    # In a production environment, this would retrieve from Key Vault
    # For now, we'll simulate with mock data
    
    # Comment the following for production implementation
    # This is just a development placeholder
    return {
        "accessKey": f"VESSEL_{vessel_id}_ACCESS_KEY",
        "secretKey": f"VESSEL_{vessel_id}_SECRET_KEY",
        "bucket": "revelare-vessel-data",
    }
    
    # Uncomment for production implementation with Key Vault
    """
    try:
        # Get Key Vault connection information from app settings
        key_vault_name = os.environ["KEY_VAULT_NAME"]
        key_vault_uri = f"https://{key_vault_name}.vault.azure.net"
        
        # Use managed identity to authenticate to Key Vault
        credential = DefaultAzureCredential()
        secret_client = SecretClient(vault_url=key_vault_uri, credential=credential)
        
        # Retrieve secrets for this vessel
        # The naming convention for secrets should be structured:
        # {vessel-id}-aws-access-key, {vessel-id}-aws-secret-key, etc.
        access_key = secret_client.get_secret(f"{vessel_id}-aws-access-key").value
        secret_key = secret_client.get_secret(f"{vessel_id}-aws-secret-key").value
        bucket = secret_client.get_secret(f"{vessel_id}-aws-bucket").value
        
        return {
            "accessKey": access_key,
            "secretKey": secret_key,
            "bucket": bucket
        }
    except Exception as e:
        logging.error(f"Error retrieving credentials from Key Vault: {str(e)}")
        raise
    """

def generate_aws_signature(access_key, secret_key, request_details):
    """
    Generates AWS Signature Version 4 signing keys and signature.
    
    This is a placeholder for the actual AWS signature generation logic
    that would be implemented in production.
    
    Parameters:
        access_key: AWS access key ID
        secret_key: AWS secret access key
        request_details: Dict containing request details for signature
        
    Returns:
        Dict with signature and authorization header
    """
    # This would contain the full AWS Signature V4 implementation
    # For brevity, showing just the concept
    
    # In production, this would implement the complete AWS Signature Version 4 algorithm
    # as documented at: 
    # https://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-authenticating-requests.html
    
    return "AWS4-HMAC-SHA256 Credential=..."