src/functions/AWSAuthenticationFunction/requirements.txt
azure-functions
azure-keyvault-secrets>=4.0.0
azure-identity>=1.4.0